docs: add SDS-Repair extension (#176)
Added SDS-R, an (optional) extension of SDS to allow for coordinated
repair of missing messages over a limited time window.

It functions by allowing subgroups of participants to rebroadcast
dependencies that were reported missing by other participants. As with
the rest of SDS, it aims to scale to larger groups first and could
likely be simplified for 1:1 or small group chats. To prevent an
explosion in repair requests or broadcast storms, it uses backoff timers
to prevent multiple participants from performing the same action (either
request a repair or rebroadcast in response to a repair request) in a
probabilistic manner.

Note that what is still missing is a specified recommended way to use
SDS-R in conjunction with retrieving missing dependencies from Store
nodes.

---------

Co-authored-by: fryorcraken <110212804+fryorcraken@users.noreply.github.com>